Relocate query_filters code for refactoring
